CRC: move TestFlag related functions to seperate test modules
